refactor(*): move `UnknownException` to global scope
