+++ src/main/java/sleppynavigators/studyupbackend/exception/ErrorCode.java
@@ -20,7 +20,9 @@ public enum ErrorCode {
 
     // business exceptions
     SESSION_EXPIRED(400, "B40", "Session expired"),
-    INTERNAL_SERVER_ERROR(500, "B50", "Internal server error"),
+
+    // global exceptions
+    INTERNAL_SERVER_ERROR(500, "G50", "Internal server error"),
     ;
 
     private final int status;
